requirements.txt
google-adk
python-dotenv
sqlalchemy
chromadb
google-generativeai
passlib
bcrypt==4.0.1
fastapi
uvicorn
itsdangerous
